# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stitching Production ERP - Fabric Store Module API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Article {
  id: number;
  articleCode: string;
  articleName: string;
  /** @nullable */
  collectionName?: string | null;
  /** @nullable */
  brandCustomer?: string | null;
  fabricType: string;
  season: string;
  category: string;
  isActive: boolean;
  /** @nullable */
  totalFabricPerUnit?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArticleBody {
  articleCode: string;
  articleName: string;
  /** @nullable */
  collectionName?: string | null;
  /** @nullable */
  brandCustomer?: string | null;
  fabricType: string;
  season: string;
  category: string;
}

export interface UpdateArticleBody {
  articleCode?: string;
  articleName?: string;
  /** @nullable */
  collectionName?: string | null;
  /** @nullable */
  brandCustomer?: string | null;
  fabricType?: string;
  season?: string;
  category?: string;
}

export interface ArticleComponent {
  id: number;
  articleId: number;
  componentName: string;
  componentType: string;
  /** @nullable */
  fabricType?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  designPrint?: string | null;
  requiredMeters: number;
  unitType: string;
  /** @nullable */
  wastagePercent?: number | null;
  createdAt: string;
}

export interface ArticleWithComponents {
  id: number;
  articleCode: string;
  articleName: string;
  /** @nullable */
  collectionName?: string | null;
  /** @nullable */
  brandCustomer?: string | null;
  fabricType: string;
  season: string;
  category: string;
  isActive: boolean;
  /** @nullable */
  totalFabricPerUnit?: number | null;
  createdAt: string;
  updatedAt: string;
  components: ArticleComponent[];
}

export interface CreateComponentBody {
  componentName: string;
  componentType: string;
  /** @nullable */
  fabricType?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  designPrint?: string | null;
  requiredMeters: number;
  unitType: string;
  /** @nullable */
  wastagePercent?: number | null;
}

export interface UpdateComponentBody {
  componentName?: string;
  componentType?: string;
  /** @nullable */
  fabricType?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  designPrint?: string | null;
  requiredMeters?: number;
  unitType?: string;
  /** @nullable */
  wastagePercent?: number | null;
}

export interface BulkComponentsBody {
  components: CreateComponentBody[];
}

export interface ComponentTemplate {
  id: number;
  templateName: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface TemplateItem {
  id: number;
  componentName: string;
  componentType: string;
  /** @nullable */
  fabricType?: string | null;
  requiredMeters: number;
  unitType: string;
  /** @nullable */
  wastagePercent?: number | null;
}

export interface ComponentTemplateWithItems {
  id: number;
  templateName: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  items: TemplateItem[];
}

export interface CreateTemplateItemBody {
  componentName: string;
  componentType: string;
  /** @nullable */
  fabricType?: string | null;
  requiredMeters: number;
  unitType: string;
  /** @nullable */
  wastagePercent?: number | null;
}

export interface CreateTemplateBody {
  templateName: string;
  /** @nullable */
  description?: string | null;
  items: CreateTemplateItemBody[];
}

export interface GrnEntry {
  id: number;
  grnNumber: string;
  date: string;
  supplierName: string;
  articleId: number;
  /** @nullable */
  articleCode?: string | null;
  /** @nullable */
  articleName?: string | null;
  totalRolls: number;
  totalMeters: number;
  ratePerMeter: number;
  totalCost: number;
  /** @nullable */
  batchNumber?: string | null;
  /** @nullable */
  colorLot?: string | null;
  /** @nullable */
  qualityType?: string | null;
  /** @nullable */
  rackLocation?: string | null;
  createdAt: string;
}

export interface CreateGrnBody {
  grnNumber: string;
  date: string;
  supplierName: string;
  articleId: number;
  totalRolls: number;
  totalMeters: number;
  ratePerMeter: number;
  /** @nullable */
  batchNumber?: string | null;
  /** @nullable */
  colorLot?: string | null;
  /** @nullable */
  qualityType?: string | null;
  /** @nullable */
  rackLocation?: string | null;
}

export interface InventorySummaryItem {
  articleId: number;
  articleCode: string;
  articleName: string;
  fabricType: string;
  totalMetersReceived: number;
  totalRolls: number;
  totalCost: number;
  grnCount: number;
}

export interface LowStockAlert {
  articleId: number;
  articleCode: string;
  articleName: string;
  totalMetersReceived: number;
  threshold: number;
}

export interface ArticleStockDetail {
  articleId: number;
  articleCode: string;
  articleName: string;
  fabricType: string;
  totalMetersReceived: number;
  totalRolls: number;
  totalCost: number;
  grnEntries: GrnEntry[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardSummary {
  totalArticles: number;
  activeArticles: number;
  totalFabricMeters: number;
  totalGrnEntries: number;
  totalStockValue: number;
  lowStockCount: number;
  recentGrnCount: number;
  topCategories: CategoryCount[];
}

export type ActivityItemMetadata = { [key: string]: unknown };

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  metadata?: ActivityItemMetadata;
}

export interface FabricByTypeItem {
  fabricType: string;
  totalMeters: number;
  articleCount: number;
}

export type ListArticlesParams = {
  search?: string;
  fabricType?: string;
  season?: string;
  category?: string;
  collection?: string;
  isActive?: boolean;
};

export type ListGrnEntriesParams = {
  articleId?: number;
  supplier?: string;
  fromDate?: string;
  toDate?: string;
};

export type GetLowStockAlertsParams = {
  threshold?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};