# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stitching Production ERP - Fabric Store Module API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List all articles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: fabricType
          in: query
          schema:
            type: string
        - name: season
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: collection
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "409":
          description: Duplicate article code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleWithComponents"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}/toggle-active:
    patch:
      operationId: toggleArticleActive
      tags: [articles]
      summary: Toggle article active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{articleId}/components:
    get:
      operationId: listArticleComponents
      tags: [components]
      summary: List components for an article
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleComponent"
    post:
      operationId: createArticleComponent
      tags: [components]
      summary: Add a component to an article
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponentBody"
      responses:
        "201":
          description: Created component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleComponent"
  /articles/{articleId}/components/bulk:
    put:
      operationId: bulkUpdateComponents
      tags: [components]
      summary: Replace all components for an article
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkComponentsBody"
      responses:
        "200":
          description: Updated components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleComponent"
  /components/{id}:
    patch:
      operationId: updateComponent
      tags: [components]
      summary: Update a component
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComponentBody"
      responses:
        "200":
          description: Updated component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleComponent"
    delete:
      operationId: deleteComponent
      tags: [components]
      summary: Delete a component
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /grn:
    get:
      operationId: listGrnEntries
      tags: [grn]
      summary: List GRN entries
      parameters:
        - name: articleId
          in: query
          schema:
            type: integer
        - name: supplier
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of GRN entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrnEntry"
    post:
      operationId: createGrnEntry
      tags: [grn]
      summary: Create a new GRN entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrnBody"
      responses:
        "201":
          description: Created GRN entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrnEntry"
  /grn/{id}:
    get:
      operationId: getGrnEntry
      tags: [grn]
      summary: Get GRN entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GRN entry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrnEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGrnEntry
      tags: [grn]
      summary: Delete a GRN entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Article:
      type: object
      properties:
        id:
          type: integer
        articleCode:
          type: string
        articleName:
          type: string
        collectionName:
          type: ["string", "null"]
        brandCustomer:
          type: ["string", "null"]
        fabricType:
          type: string
        season:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        totalFabricPerUnit:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - articleCode
        - articleName
        - fabricType
        - season
        - category
        - isActive
        - createdAt
        - updatedAt
    CreateArticleBody:
      type: object
      properties:
        articleCode:
          type: string
        articleName:
          type: string
        collectionName:
          type: ["string", "null"]
        brandCustomer:
          type: ["string", "null"]
        fabricType:
          type: string
        season:
          type: string
        category:
          type: string
      required:
        - articleCode
        - articleName
        - fabricType
        - season
        - category
    UpdateArticleBody:
      type: object
      properties:
        articleCode:
          type: string
        articleName:
          type: string
        collectionName:
          type: ["string", "null"]
        brandCustomer:
          type: ["string", "null"]
        fabricType:
          type: string
        season:
          type: string
        category:
          type: string
    ArticleComponent:
      type: object
      properties:
        id:
          type: integer
        articleId:
          type: integer
        componentName:
          type: string
        componentType:
          type: string
        fabricType:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        designPrint:
          type: ["string", "null"]
        requiredMeters:
          type: number
        unitType:
          type: string
        wastagePercent:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - articleId
        - componentName
        - componentType
        - requiredMeters
        - unitType
        - createdAt
    ArticleWithComponents:
      type: object
      properties:
        id:
          type: integer
        articleCode:
          type: string
        articleName:
          type: string
        collectionName:
          type: ["string", "null"]
        brandCustomer:
          type: ["string", "null"]
        fabricType:
          type: string
        season:
          type: string
        category:
          type: string
        isActive:
          type: boolean
        totalFabricPerUnit:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        components:
          type: array
          items:
            $ref: "#/components/schemas/ArticleComponent"
      required:
        - id
        - articleCode
        - articleName
        - fabricType
        - season
        - category
        - isActive
        - createdAt
        - updatedAt
        - components
    CreateComponentBody:
      type: object
      properties:
        componentName:
          type: string
        componentType:
          type: string
        fabricType:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        designPrint:
          type: ["string", "null"]
        requiredMeters:
          type: number
        unitType:
          type: string
        wastagePercent:
          type: ["number", "null"]
      required:
        - componentName
        - componentType
        - requiredMeters
        - unitType
    UpdateComponentBody:
      type: object
      properties:
        componentName:
          type: string
        componentType:
          type: string
        fabricType:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        designPrint:
          type: ["string", "null"]
        requiredMeters:
          type: number
        unitType:
          type: string
        wastagePercent:
          type: ["number", "null"]
    BulkComponentsBody:
      type: object
      properties:
        components:
          type: array
          items:
            $ref: "#/components/schemas/CreateComponentBody"
      required:
        - components
...
    GrnEntry:
      type: object
      properties:
        id:
          type: integer
        grnNumber:
          type: string
        date:
          type: string
          format: date-time
        supplierName:
          type: string
        articleId:
          type: integer
        articleCode:
          type: ["string", "null"]
        articleName:
          type: ["string", "null"]
        totalRolls:
          type: integer
        totalMeters:
          type: number
        ratePerMeter:
          type: number
        totalCost:
          type: number
        batchNumber:
          type: ["string", "null"]
        colorLot:
          type: ["string", "null"]
        qualityType:
          type: ["string", "null"]
        rackLocation:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - grnNumber
        - date
        - supplierName
        - articleId
        - totalRolls
        - totalMeters
        - ratePerMeter
        - totalCost
        - createdAt
    CreateGrnBody:
      type: object
      properties:
        grnNumber:
          type: string
        date:
          type: string
          format: date-time
        supplierName:
          type: string
        articleId:
          type: integer
        totalRolls:
          type: integer
        totalMeters:
          type: number
        ratePerMeter:
          type: number
        batchNumber:
          type: ["string", "null"]
        colorLot:
          type: ["string", "null"]
        qualityType:
          type: ["string", "null"]
        rackLocation:
          type: ["string", "null"]
      required:
        - grnNumber
        - date
        - supplierName
        - articleId
        - totalRolls
        - totalMeters
        - ratePerMeter
...